Problem: Packets larger than UDP's MTU needs to be broken up
Solution: break up a file into 1024 byte chunks.
The first response contains how many chunks the total batch
consists of. The first packet uses the same name as the request
name.
i.e.

name: PiratesoftheCaribbean content: PiratesOfTheCaribbean-10000
the name converts to vec![vec![806, 1462, 1766]]

The second packet will be named
PiratesOfTheCarribean/0
which'll translate to an sdri
vec![vec![806, 1462, 1766], vec![2, 482, 1250]]

The third packet will be named
PiratesOfTheCarribean/1
which translates to an sdri
vec![vec![806, 1462, 1766], vec![71, 1415, 1687]]

So this way the initial search for a packet will hit a
face with the correct breadcrumb.
Thereafter we can query for the union of the sequence and name

Importantly, the sequence needs to have the <name>-sequence

If we don't do this then every sequence packet will poison every face
so sequences need uniqueness