Problem: identity shouldn't change name
Solution: detach trusted_connections hash from identity.
We should be able to construct the trusted_connections packet
from the name of the identity. The trusted_connections file
will accrete and this will be reflected in a growing packet size
with an associated name change (incrementing upon every addition
of trusted identity).