Problem: a router must have more than udp as transport
Solution: this is the 1st of a few patches that introduces
different channels into copernica. I'm not sure how it'll
turn out as this is the first of a basic structural changes
Packet transferring hasn't been implemented yet.

This patch also make huge refactorings to the codebase stripping
out anything that isn't neccessary in the copernica package so
that the eventual goal of no_std can be achieve (in the copernica
package)