Problem: Identity isn't weaved in yet.
Solution: Now that the basic routing and packet flow is working
it's about time to start layering in identity, signatures and
encryption. This is required for more serious higher level apps.

This patch lays the foundation for what a more final version of
the basic packet structure looks like. Encryption needs to be tested
and associated encryption apps firmed up.