Problem: No basic way to flash hello world onto dongle
Solution: provide two ways

1) ccli which'll be the program that'll flash the .elf to the dongle
for the user after undergoing the shamir secret sharing process

2) for development of the firmware the programmer may simply run
cargo run, which'll call probe-run in the background, automatically
downloading the artefact to the device and running it.

This way I can have two terminals open, one which'll be copernica
executable side that talk directly to the USB dongle after the flash.

Trying to make it as frictionless to get this setup working.