Problem: offset and total in the NarrowWaistPacket::Response were an incorrect assumption
Problem: the incorrect assumption was that I could issue a frame range query in one NarrowWaistPacket::Request and I'd get a series of NarrowWaistPacket::Responses...
This approach made for a living hell in the router, I could see complexity going straight through the roof and decided it was better to issue one Request and get one Response.

Solution: remove total and offset from NarrowWaistPacket::Response.